grepx-task-generator-server/requirements.txt
# Task Generator Server Requirements
# ===================================

# Core dependencies
sqlalchemy==2.0.23
pyyaml==6.0.2

# Dagster (for ConfigurableResource base class)
dagster==1.9.13

